API

API

List of functions and datastructure available in AutoViz

AutoViz.ArrowCar — Type.

ArrowCar{A<:AbstractArray{Float64}, C<:Colorant} <: Renderable
ArrowCar(pos::AbstractArray, angle::Float64=0.0; length = 4.8, width = 1.8,  color=_colortheme["COLOR_CAR_OTHER"], text="", id=0)
ArrowCar(x::Real, y::Real, angle::Float64=0.0; length = 4.8, width = 1.8,  color=_colortheme["COLOR_CAR_OTHER"], text="", id=0)

A renderable type to represent a rectangular car with an arrow in the middle.

source

AutoViz.BlinkerOverlay — Type.

BlinkerOverlay

Displays a circle on one of the top corner of a vehicle to symbolize a blinker. fields:

on: turn the blinker on

right: blinker on the top right corner, if false, blinker on the left

veh: the vehicle for which to display the blinker

color: the color of the blinker

size: the size of the blinker

source

AutoViz.HistogramOverlay — Type.

HistogramOverlay

Display a bar at the specified position pos, the bar is of size width, height and is filled up to a given proportion of its height. The fill proportion is set using val, it should be a number between 0 and 1. If it is 0, the bar is not filled, if it is 1 it is filled to the top.

source

AutoViz.IDOverlay — Type.

IDOverlay

Display the ID on top of each entity in a scene.

Fiels

color::Colorant default white

font_size::Int64 default 15

source

AutoViz.SceneFollowCamera — Type.

SceneFollowCamera{R<:Real}

Camera centered over all vehicles The zoom can be adjusted.

Fields

zoom::R

source

AutoViz.isrenderable — Function.

Return true if an object or type is directly renderable, false otherwise.

New types should implement the isrenderable(t::Type{NewType}) method.

source

AutoViz.render — Method.

render(scene)
render(scene; kwargs...)

Render all the items in scene to a Cairo surface and return it.

scene is simply an iterable object (e.g. a vector) of items that are either directly renderable or renderable by conversion. See the AutoViz README for more details.

source

AutoViz.set_color_theme — Method.

set_color_theme(colortheme)

Change the color theme of the package

source

AutoViz.write_to_svg — Method.

write_to_svg(surface::CairoSurface, filename::AbstractString)

Write a cairo svg surface to a file. The surface object is destroyed after.

source

AutoViz.renderfun — Method.

renderfun(veh::Entity{S,D,I}) where {S,D,I}

decides which rendering function to use based on the class of veh and AutoViz RENDER_MODE

source